+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 * Add setting to unlock A20 for all characters
 * Read and display StS build distributor
 * Stop the Steam overlay running on the MTS window
+* Fix `optional` not working for `clz` in SpirePatch
 
 #### v3.30.3 ####
 * Fix package information being lost for patched classes (Alchyr)